Initial commit of code
cw-kbd is a USB keyboard (HID) that you can use
to type on your computer using morse code on the
paddles.

The project is getting too hairy to exist outside
a git repo.  It is a semi-working creation that
depends on the LUFA library for USB and used the
virtual serial port mouse demo as an example of
how to set stuff up.

Signed-off-by: Vernon Mauery <vernon@mauery.com>