Use stdlib's utoa instead of custom itoa
stdlib's version of utoa is smaller (and likely more
optimized) than the version I wrote here.  We just had
to pull out the radix code to make it work.  Also the
arguments are in a different order.  Since it is not
really a standard function, this is a common problem.

Signed-off-by: Vernon Mauery <vernon@mauery.com>