convert sprintf into snprint
sprintf has some dangers when you don't always know the
maximum string length.  snprintf takes care of that.  This
converts our sprintf into snprintf, but renames it to
my_snprintf because we use a uint8_t instead of a size_t
for the string length.  Given the RAM restrictions, I think
this is a reasonable change, so we had to rename in order
to avoid a collision.

Signed-off-by: Vernon Mauery <vernon@mauery.com>