Add some string names for cw state machine enums
Since it is so hard to translate state numbers into anything
useful, this adds some arrays that have string values of the
enums in them.  To add arrays of progmem strings, you have to
declare the strings and then add progmem arrays of the progmem
strings.  It is more than a little obnoxious.  But it saves a
load of space that would otherwise be in RAM.  And it gives me
debug prints like:
cw_in: state keying_right_ready (keying_x_right_key_release)

Signed-off-by: Vernon Mauery <vernon@mauery.com>