cw timing fixes (in and out state machines)
On the in machine, make it so on a key release from the
both_pressed state, you only need to wait DIT cycles before
enqueuing the next didah.  This also removes a conditional
assignment.

On the out machine, if you are dequeuing didahs and the
queue dries up, go back to idle.  Only enqueue spaces if
a space is in the queue.

Signed-off-by: Vernon Mauery <vernon@mauery.com>