Fix back-to-back enqueued characters
If a hid report sends two back-to-back reports with the same
key pressed, it will just assume that the key was pressed
continuously for those two reports, thus not causing a new
event to get emitted.  This causes the /AA prosign to get
emitted as /A.  This commit enforces the no-2-back-to-back
character rule.  If the next character is the same as the
last one, it will send an empty report (no keys pressed) and
wait for the next report to continue emptying the queue.

Signed-off-by: Vernon Mauery <vernon@mauery.com>