A newer better command mode
Command mode was getting a little bit hairy and this makes it
ten times worse.  Well, not really.  It does make it considerably
bigger, but it does a better job of keeping all the control logic
separate from the i/o logic.  This makes it easier to understand
the complex state machine.  With the added benefit of simplified
complexity, this commit adds a little bit more feedback on what
is going on so the user won't get so lost.

Signed-off-by: Vernon Mauery <vernon@mauery.com>