Add cootie (bug) and paddle modes
I am not entirely sure if this is what that mode means.  I really
need an elmer.  As far as I can gather, cootie mode is a semi-
automatic mode where it sends dits as long as you press dit and
dahs as long as you press dah.  If one is pressed while holding
the other, the second will sound until released, but the first
will not continue to sound.  This behavior is much like a computer
keyboard.

And from what I understand paddle mode is where you can only press
one key at a time, but you only get one dit or dah out of each key
so it is a non-automatic bug.

Signed-off-by: Vernon Mauery <vernon@mauery.com>