Add a memory auto-repeat function
The memory auto-repeat function allows the user to program the
device to play a message every N minutes, where N < 100.  This
changes the eeprom layout, so beware.  While this change goes in
it reinforces how silly it is to have a separate 'callsign' memory
location.  This commit also removes that in favor of a 'hot-button'
memory location 0 that is accessible via 'e' in command mode.

Signed-off-by: Vernon Mauery <vernon@mauery.com>