Add a reset button
The command mode button, if pressed for 10 seconds
will act as a reset to change all eeprom settings
back to defaults.

At this point in time, this also requires a reboot
of the device to re-read the new settings.  We may
in the future do this via software, but for now, it
must be done physically.

Signed-off-by: Vernon Mauery <vernon@mauery.com>