Change debug from a macro to a function
Debugging statements were taking up too much space with the
inline macro.  This commit turns debug into a function.  As
a side-effect, it also exposes a new snvprintf interface that
has the vargs version of snprintf.  Really what it does is
make the snprintf a wrapper to snvprintf.  But I don't think
it matters because snprintf isn't used outside of debug.

Signed-off-by: Vernon Mauery <vernon@mauery.com>