Add spaces into the queue properly
In the in machine, dits and dahs are timed such that
each has an inter-didah space attached.  Inter-character
space is required or the decoder will not work.  So all
inter-character spaces get inserted when the keyer is idle
and show up as one space mark in the decode queue.  Inter-
word spaces show up as two spaces in the decode queue, but
only after a six-dit idle time (remember the inter-didah
space makes seven).

This also changes the decode method to NOT add spaces after
non-printing-characters so blunders don't end up as a long
string of spaces.

Signed-off-by: Vernon Mauery <vernon@mauery.com>