Fix iambic keyer
The iambic keyer was not behaving correctly. It added one more
of the key that was released, but it should just add the next
of what is coming.  Plus, it had a strange glitch because it
was adding the didahs at the wrong place.

This commit adds a didah immediately and then starts counting
from there.

Signed-off-by: Vernon Mauery <vernon@mauery.com>