Move usb work into SOF event
USB has a start of frame event that gives us a fairly precise
1ms timer.  I would just turn off usb_work when USB is not
alive, but you can only detect if usb is not alive if you
cut the power contact and make it so the teensy is not bus
powered.  This way, we can only do usb_work when there are
usb packets going.  This paves the way for the tickless core.

Signed-off-by: Vernon Mauery <vernon@mauery.com>