Settings profiles are now complete
All the settings except for the memory and repeat
settings can be saved in profiles.  Whatever the
current profile is gets any new settings applied
to it.  The you can switch between profiles and
all your settings migrate.  This allows for several
people to use the same keyer or to have different
settings for practicing other styles.  It also
includes a method by which you can reset the
current profile to sane defaults.

Signed-off-by: Vernon Mauery <vernon@mauery.com>