Debouncing int6 in a more efficient manner
INT6 is only a system interrupt that can wake from sleep
mode 2 when it is a low-level triggered interrupt.  But
this causes problems with the old method of debouncing
because if the button is held down, it issues a constant
stream of interrupts.  This new method disables the
interrupt as soon as we get the first one and then starts
a ms_tick function that counts how many milliseconds the
button is held down for.

o At 10ms, the button is considered 'pressed' and enters
  command mode
o At 2000ms, the button is 'held' and sets current profile
  back to sane defaults
o At 5000ms, the button is a reset button that clears all
  the profiles and memory settings

Signed-off-by: Vernon Mauery <vernon@mauery.com>