Another try at soft reset
This seems to be better, but I have still seen it hang
sometimes.  But I have also seen it work like this, so
I am not sure where the problem lies.  I have grand plans
to extend a reset button out the bottom of the case for
emergencies when the soft reset doesn't work.

Signed-off-by: Vernon Mauery <vernon@mauery.com>